Primary setup utility for stpR.

This function is used to create the initial metadata objects that are used by the other setup utilities in stpR. Processes stp_ob artifacts for multiple scenarios:

Usage

eval_map(
  data_df = NULL,
  from = NULL,
  to = NULL,
  std_proc_na = NULL,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id = NULL,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here(),
  summary = FALSE
)

Arguments

data_df

The data frame to be evaluated.

from

string value of the column(s) in the data frame that represent the columns being processed in the current data set

to

string value of the column(s) in the data frame that represent the columns being processed in the current data set.

std_proc_na

A function that will be applied to the from column(s) to determine which values should be considered "standard processing" NA values. This field is used in cases when processing values such as numbers, where it's not necessary to know all the "typically" processed values, but it is helpful to know the values that aren't readily processed by this standard processing as well as how they are handled in the current data set.

highlight

A boolean tag to denote whether this processing is of interest for highlighting in the reporting of the current state of the data. Generally, this would be a type of mapping or field processing that would be of interest to routinely review for each data set processing.

issue

A boolean tag to denote whether this processing is of interest for noting for inspection as an issue in the reporting of the current state of the data.

notes

A string value to be included in the reporting of the current state of the data.

stp_id

A string value to be used as the identifier for this processing in the reporting of the current state of the data. By default, this is the to column.

save_artifact

A boolean tag to denote whether this processing should be saved as an artifact. Only false if you just wanted to use this function to do the count() type check with benefit of console output control.

report

A boolean tag to denote whether this processing should be included in the reporting of the current state of the data. May be false in scenarios where you didn't want to routinely look at the output for processing, but did want to be alerted when there were updates of values.

perform_compare

A boolean tag to denote whether this processing should be compared against the previous data artifact. to be used in the comparison of the current data against previous data artifacts. This should be true when you want to "enforce" a given relationship

project_dictionary

The project_dictionary object, which by default is output by the get_project_dictionary() function.

project_directory

The project_directory object, which by default is output by the here::here() function.

summary

A boolean tag to denote whether this processing is a summary of the data rather than a mapping of values. Primarily only true when this function is called from eval_summary().

Value

a data frame with the current mapping and any updates to the mapping, if relevant.

Details

Initial processing of raw data for both categorical and "continuous" fields (depending on whether std_proc_na is NULL)

General summary of value combinations for several columns: As a utility, eval_summary() also utilizes the functionality of this utility, as well: When summary = TRUE, the data_df is assumed to be a summary table rather than a processing of from and to values as is commonly the case for the use of eval_map().

Comparison of current data against previous data artifacts

Additional creation of tags of items to note as "issues" or "highlights" for reporting of the current state of the data

Examples

# \donttest{
  # mapping with issue
  eval_map(
    data_df = qa_tbl,
    from = "TRT_RAW",
    to = "trt",
    notes = "Note multiple raw processed values associated with each treatment type"
  )
#> Error in loadNamespace(x): there is no package called ‘here’

  qa_tbl$age <- as.numeric(qa_tbl$AGE_RAW)
#> Warning: NAs introduced by coercion
  qa_tbl$age[qa_tbl$AGE_RAW == "twenty"] <- 20

  # 'continuous' field with issue
  eval_map(
    data_df = qa_tbl,
    from = "AGE_RAW",
    to = "age",
    std_proc_na = as.numeric,
    notes = "Note processing of value of 'twenty' from AGE_RAW"
  )
#> Error in loadNamespace(x): there is no package called ‘here’

# }