Create artifact for summary cases.

For use with stpR setup cases where a pre-defined mapping of values is of interest to be tracked, but where variables are strictly associated in a from and to relationship. Most likely to be utilized

Usage

eval_summary(
  summary_df = NULL,
  key,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)

Arguments

summary_df

The data frame, pre-summarized, to be associated with this documentation item.

key

String value of the column name to be used as the key for the summary_df. Will be utilized similarlly to the from column in the eval_map() function. In some scenarios, deciding on the key variable will be arbitrary, but choosing it here to support additional comparison functionality associated with these artifact items.

highlight

A boolean tag to denote whether this processing is of interest for highlighting in the reporting of the current state of the data. Generally, this would be a type of mapping or field processing that would be of interest to routinely review for each data set processing.

issue

A boolean tag to denote whether this processing is of interest for noting for inspection as an issue in the reporting of the current state of the data.

notes

A string value to be included in the reporting of the current state of the data.

stp_id

A string value to be used as the identifier for this processing in the reporting of the current state of the data. By default, this is the to column.

save_artifact

A boolean tag to denote whether this processing should be saved as an artifact. Only false if you just wanted to use this function to do the count() type check with benefit of console output control.

report

A boolean tag to denote whether this processing should be included in the reporting of the current state of the data. May be false in scenarios where you didn't want to routinely look at the output for processing, but did want to be alerted when there were updates of values.

perform_compare

A boolean tag to denote whether this processing should be compared against the previous data artifact. to be used in the comparison of the current data against previous data artifacts. This should be true when you want to "enforce" a given relationship

project_dictionary

The project_dictionary object, which by default is output by the get_project_dictionary() function.

project_directory

The project_directory object, which by default is output by the here::here() function.

Value

a data frame with the current mapping and any updates to the mapping, if relevant.

Details

This function calls eval_map() with the summary parameter set to TRUE. See eval_map() for additional details.

Examples

# \donttest{
  eval_summary(
    summary_df = count(qa_tbl, approach, asa),
    key = "proc_name",
    notes = "summary proc, ASA notes",
    stp_id = "comp_approach_asa"
  )
#> Error in count(qa_tbl, approach, asa): could not find function "count"

# }