Primary utility for logging items to text metadata items for the stpR package.

Logs assert failures, comments, and todos for review. Both save_metadata_global and include_metadata_log_items must be set to TRUE in the project_dictionary object for items to be logged. Expected that during initial setup steps, may want to avoid logging all log_items values. To avoid overhead of maintaining meaningful, individual ids for each log item, the default stp_id is set to "99999". All instances will be logged (including duplicate items, etc.). Function will be created to remove individual or all log items. For a final run of this documentation per iteratio of data set processed.

Usage

log_item(
  type,
  item,
  add_item = NULL,
  priority = NULL,
  stp_id = "99999",
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)

Arguments

type

String value indicating type of item to log. Either "ALERT", "TODO", or "COMMENT".

item

String value of the item to log.

add_item

Additional item to log. This is used for the assert_distinct, assert_cases, assert_order output.

priority

String value indicating priority of item to log. TBD if this is still a useful field to include

stp_id

Id of the stp_ob entity. Defaults to "99999". Not as significant in usage as with other strR logging functions, such as eval_map(), etc.

project_dictionary

The project_dictionary object, which by default is output by the get_project_dictionary() function.

project_directory

The project_directory object, which by default is output by the here::here() function.

Examples

# \donttest{
# Example 1 ----------------------------------
log_item(
  type = "TODO",
  item = "Create function to reset all log_item metadata",
  priority = "1"
 )
#> Error in loadNamespace(x): there is no package called ‘here’

# Example 2 ----------------------------------
# Called by assert_cases.R function
log_item(type = "ALERT",
    item = eval_logic_assert_string,
    priority = "1",
 )
#> Error in loadNamespace(x): there is no package called ‘here’
# }